+++ js/data.js
@@ -1,0 +1,24 @@
+const DESCRIPTIONS = [
+  'Отличный день на море!', 'Закат в горах просто волшебный', 'Кофе утром - лучший старт дня',
+  'Прогулка по осеннему парку', 'Мой пушистый друг', 'Городские огни ночью', 'Сказочное Бали',
+  'Вкусный ужин в хорошей компании', 'Уютный вечер с книгой', 'Природа в ее лучшем проявлении',
+  'Архитектурный шедевр', 'Мои спортивные достижения', 'Творческий процесс', 'Семейные моменты',
+  'Музыка вдохновляет', 'Уличное искусство', 'Кулинарные эксперименты', 'Рассвет нового дня',
+  'Исторические места', 'Технологии будущего', 'Романтический вечер', 'Рабочие будни',
+  'Отдых на природе', 'Городская жизнь', 'Редкие моменты счастья после получения стипухи'
+];
+
+const MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+
+const NAMES = [
+  'Иван', 'Мария', 'Алексей', 'Элой', 'Давид', 'Анна', 'Т-34', 'Матильда','Святозар', 'Коломбина', 'Михаил', 'Виктория', 'Андрей', 'Григорий', 'Зена королева воинов'
+];
+
+export { DESCRIPTIONS, MESSAGES, NAMES };
